feat: implement cluster node architecture with subnode providers
refactor(uuid): replace new_v4 with now_v7 for time-ordered UUIDs
style: normalize formatting in utils and examples
fix(email_trigger): improve trigger_id generation safety
docs: add cluster node architecture and deepseek provider guides
test: add comprehensive tests for deepseek provider

+++ hetumind/hetumind-nodes/src/llm/shared.rs
@@ -192,7 +192,7 @@ pub struct ModelCapabilities {
   pub streaming: bool,
   pub function_calling: bool,
   pub vision: bool,
-  pub max_context_length: Option<u32>,
+  pub max_context_length: Option<u64>,
   pub supported_formats: Vec<String>,
   pub json_mode: bool,
   pub system_messages: bool,
